% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastInternals.R
\name{blastnw}
\alias{blastnw}
\title{Seach with Blatns for Windows.}
\usage{
blastnw(genome = "", db = "", outdir = ".", n_threads = 1L, eval = 1e-06)
}
\arguments{
\item{genome}{The query. A nucleotide fasta file.}

\item{db}{The subject. A nucleotide blast database as created by
\link{makeblastdb}.}

\item{outdir}{Where to put the blast output file.}

\item{n_threads}{\code{integer}. The nomber of cores to use.}

\item{eval}{Evalue reporting threshold.}
}
\value{
The path to the blastn output. The blastn output is written in the
following format as specified by \code{-outfmt} option in Blast+ program:
\code{-outfmt '6 qseqid sseqid pident gaps length slen evalue qseq'}.
}
\description{
Seach nucleotide sequences in a nucleotide database.
}
\author{
Alba Arranz García
}
