% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickMLST.R
\name{doMLST}
\alias{doMLST}
\title{Perform MLST Analysis Over a List of Genomes}
\usage{
doMLST(infiles, org = "leptospira", scheme = 1L, schemeFastas = NULL,
  schemeProfile = NULL, write = "new", ddir = paste0("pubmlst", "_", org,
  "_", scheme), fdir = paste0("alleles", "_", org, "_", scheme),
  n_threads = 1L, pid = 90L, scov = 0.9)
}
\arguments{
\item{infiles}{A vector of genome sequences in fasta format.}

\item{org}{A valid organism from pubmlst.org. Run \link{listPubmlst_orgs} to
see available ones.}

\item{scheme}{\code{integer}. The scheme id number for a given organism. Run
\link{listPubmlst_schemes} to see available schemes for certain \code{org}.}

\item{schemeFastas}{A vector with the path to fasta sequences from each
loci of the specified mlst scheme. If it is \code{NULL} (default), sequences
are downloaded from \url{http://rest.pubmlst.org} to \code{dir}.}

\item{schemeProfile}{The path to the profile file (.tab). If left \code{NULL}
then it is downloaded from \url{http://rest.pubmlst.org} to \code{dir}.}

\item{write}{\code{character}. One of \code{"new"} (Default), \code{"all"} or
\code{"none"}. The fist one writes only new alleles found (not reported in
pubmlst.org), the second writes all alleles found, and "none" do not write
any file.}

\item{ddir}{A non-existing directory where to download the loci fasta files
in case they are not provided by the user. Default:
\code{paste0('pubmlst','_',org,'_',scheme)}}

\item{fdir}{A non-existing directory where to write fasta files of found
sequences (see \code{write}). (Default:
\code{paste0('alleles','_',org,'_',scheme)}).}

\item{n_threads}{\code{integer}. The number of cores to use. Each job consist
on the process for one genome. Blastn searches will use 1 core per job.}

\item{pid}{Percentage identity threshold to be consider as an allele. An
\code{integer} <= 100. (Default: 90).}

\item{scov}{Subject coverage threshold to be consider as an allele. A
\code{numeric} between 0 and 1. Not recomended to set it below 0.7 .
(Default 0.9)
# @details}
}
\value{
A \code{data.frame}. Rows are \code{infiles} and columns are genes
from the selected MLST scheme. '<NA>' values means that no allele were
found in the respective genome. 'u' (from 'unknown') means that the allele
found was not reported in pubmlst.org database; a fasta file with the new
allele is written in this case. A number indicates the allele id number of
the reported alleles in pubmlst.org . The last column ($ST) indicates the ST
determined by the sequence of alleles, if available.
}
\description{
Takes a list of genome fasta files and perform blastn searches
to identify the sequence type for each of the genes/loci available in a mlst
scheme.
}
\references{
Altschul, Gish, Miller, Myers & Lipman. (1990) "Basic local
alignment search tool." \emph{J. Mol. Biol}. \strong{215}:403-410.

Jolley KA, Bray JE, Maiden MCJ. A RESTful application programming interface
for the PubMLST molecular typing and genome databases. Database: The Journal
of Biological Databases and Curation. 2017;2017:bax060.
doi:10.1093/database/bax060.
}
\author{
Ignacio Ferres
}
