% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmlstInterface.R
\name{downloadPubmlst_profile}
\alias{downloadPubmlst_profile}
\title{Download the Profile for a Given MLST Scheme}
\usage{
downloadPubmlst_profile(org = "leptospira", scheme = 1L, dir = ".")
}
\arguments{
\item{org}{\code{character}. A genus (lowercase).}

\item{scheme}{An \code{integer} specifiying the scheme number.}

\item{dir}{Where to put the tsv profile file.}
}
\value{
A vector of paths to the downloaded tsv file.
}
\description{
This function takes a genus, as returned by \link{listPubmlst_orgs},
and a mlst scheme as an \code{integer} and downloads the profile file for
the given genus and mlst scheme, in tsv format.
}
\references{
Jolley KA, Bray JE, Maiden MCJ. A RESTful application programming
interface for the PubMLST molecular typing and genome databases. Database:
The Journal of Biological Databases and Curation. 2017;2017:bax060.
doi:10.1093/database/bax060.
}
\author{
Ignacio Ferres
}
