% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getf}
\alias{getf}
\title{Get a vector of integers which represents a factor to split the
sequences in equal alleles taking into account the reverse complement.}
\usage{
getf(sq)
}
\arguments{
\item{sq}{A list with SeqFastadna objects.}
}
\value{
A vector of integers
}
\description{
Get a vector of integers which represents a factor to split the
sequences in equal alleles taking into account the reverse complement.
}
