% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastInternals.R
\name{makeblastdb}
\alias{makeblastdb}
\title{Make a Blast Database}
\usage{
makeblastdb(infile, hash_index = F, parse_seqids = F)
}
\arguments{
\item{infile}{A nucleotide fasta file.}

\item{hash_index}{\code{logical}. See Blast+ manual.}

\item{parse_seqids}{\code{logical}. See Blast+ manual.}
}
\value{
The path to the created database.
}
\description{
Takes a nucleotide fasta file and produce a blast database.
}
\references{
Altschul, Gish, Miller, Myers & Lipman. (1990) "Basic local
 alignment search tool." \emph{J. Mol. Biol}. \strong{215}:403-410.
}
\author{
Ignacio Ferres
}
