% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickMLST.R
\name{mlst}
\alias{mlst}
\title{MLST}
\usage{
mlst(genome, dbs, write.new = TRUE, dir = ".", n_threads = 1L,
  outf = tempdir(), pid = 90, scov = 0.9)
}
\arguments{
\item{genome}{A FASTA file with the genome sequence.}

\item{dbs}{Subject database files. No subfix needed.}

\item{write.new}{\code{logical}. If new alleles found should be written in
\code{dir}.}

\item{dir}{Sequences of new allele found will be placed here.}

\item{n_threads}{\code{integer}. The number of threads to use by BLASTN.}

\item{outf}{Where the blastn output will be written.}

\item{pid}{Percentage identity.}

\item{scov}{Query coverage.}
}
\value{
A vector with the mlst.
}
\description{
Perform Multi Locus Sequence Typing.
}
\author{
Ignacio Ferres
}
