% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mlstw}
\alias{mlstw}
\title{MLST}
\usage{
mlstw(
  genome,
  dbs,
  write = "new",
  prefix = "allele",
  dir = ".",
  dnw = paste0(dir, "tmp/"),
  n_threads = 1L,
  outf = tempdir(),
  pid = 90,
  scov = 0.9
)
}
\arguments{
\item{genome}{A FASTA file with the genome sequence.}

\item{dbs}{Subject database files. No subfix needed.}

\item{write}{\code{character}. One of \code{"new"} (Default), \code{"all"} or
\code{"none"}. The fist one writes only new alleles found (not reported in
pubmlst.org), the second writes all alleles found, and "none" do not write
any file.}

\item{prefix}{\code{character} A prefix to the fasta files of found
sequences (see \code{write}). (Default: \code{"allele"}).}

\item{dir}{An existing directory where to put the loci fasta files in case
they are not provided by the user. Also sequences found will be placed here
if \code{write} is set to ethier \code{"new"} or \code{"all"}.}

\item{dnw}{Temporary directory where to put new alleles to check later if
are the same.}

\item{n_threads}{\code{integer}. The number of threads to use by BLASTN.}

\item{outf}{Where the blastn output will be written.}

\item{pid}{Percentage identity.}

\item{scov}{Query coverage.}
}
\value{
A vector with the mlst.
}
\description{
Perform Multi Locus Sequence Typing in windows
}
\author{
Alba Arranz García
}
