% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_mlst.R
\name{plot.mlst}
\alias{plot.mlst}
\title{Plot A mlst Object}
\usage{
\method{plot}{mlst}(mlst, what = "both", vertex.size = 3,
  vertex.label = NA, layout = layout.fruchterman.reingold, plot = TRUE,
  ...)
}
\arguments{
\item{mlst}{An object of class \code{mlst}.}

\item{what}{One of "result", "profile", "both" (default). What should be
plotted. White nodes are plotted for those isolates with no ST assigned. If
"both", isolates with assigned ST are plotted in blue.}

\item{vertex.size}{The size of the vertex. Default: 3.}

\item{vertex.label}{Default: NA.}

\item{layout}{layout. Default: layout.fruchterman.reingold.}

\item{plot}{Default: TRUE.}

\item{...}{Further arguments to pass to \link[igraph]{plot.igraph}.}
}
\value{
A minimum spanning tree plot and an object of class \code{igraph}
(invisible).
}
\description{
Plot a \code{mlst} object. A Minimum Spanning Tree is generated
and a graph plot is rendered.
}
