% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastInternals.R
\name{processBlastResult}
\alias{processBlastResult}
\title{Process Blastn Result}
\usage{
processBlastResult(blastRes, pid = 90, scov = 0.9, write = "new",
  prefix = "alleles", dir = ".", dnw = paste0(dir, "tmp/"))
}
\arguments{
\item{blastRes}{A \code{data.frame} as returned by \link{readBlastResult}.}

\item{pid}{Percentage identity reporting threshold.}

\item{scov}{Query coverage reporting threshold.}

\item{write}{\code{character}. One of \code{"new"} (Default), \code{"all"} or
\code{"none"}. The fist one writes only new alleles found (not reported in
pubmlst.org), the second writes all alleles found, and "none" do not write
any file.}

\item{prefix}{\code{character} A prefix to the fasta files of found
sequences (see \code{write}). (Default: \code{"allele"}).}

\item{dir}{The directory where to put the fasta file of allele sequences
found.}

\item{dnw}{Temporary directory where to put new alleles to check later if
are the same.}
}
\value{
The allele number id if an exact match with a reported mlst gene is
found, a name arbitrarily given if a new allele is found, or \code{NA} if no
alleles are found.
}
\description{
Process blastn output. Reports whether the genome
contains a reported allele in pubmlst, or it has a not reported (new)
allele, or if no allele where found. In case a new allele were found, a
fasta file is written.
}
\author{
Ignacio Ferres
}
