% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastInternals.R
\name{readBlastResult}
\alias{readBlastResult}
\title{Read Blast Result}
\usage{
readBlastResult(blout = "")
}
\arguments{
\item{blout}{The path to the blastn output, written in format as specified
by the Blast+ parameter \code{-outfmt '6 qseqid sseqid pident gaps length
slen evalue qseq'}.}
}
\value{
A \code{data.frame} or a \code{try-error} message if blout is empthy.
}
\description{
Read blast result
}
\author{
Ignacio Ferres
}
