% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmlstInterface.R
\name{listPubmlst_orgs}
\alias{listPubmlst_orgs}
\title{List Available Organisms in Pubmlst.org Database}
\usage{
listPubmlst_orgs()
}
\value{
A list of genus.
}
\description{
List all organisms available in pubmlst database.
}
\author{
Ignacio Ferres
}
