% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmlstInterface.R
\name{listPubmlst_schemes}
\alias{listPubmlst_schemes}
\title{List Available MLST Schemes for a Given Genus}
\usage{
listPubmlst_schemes(org = "leptospira")
}
\arguments{
\item{org}{\code{character}. A genus (lowercase).}
}
\value{
A \code{list} with genes for each available mlst scheme.
}
\description{
Takes a single genus as returned by \link{listPubmlst_orgs},
and lists the available mlst schemes for that genus in pubmlst.org .
}
\author{
Ignacio Ferres
}
