% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastInternals.R
\name{processBlastResult}
\alias{processBlastResult}
\title{Process Blastn Result}
\usage{
processBlastResult(blastRes, pid = 90, scov = 0.9, write.new = TRUE,
  dir = ".")
}
\arguments{
\item{blastRes}{A \code{data.frame} as returned by \link{readBlastResult}.}

\item{pid}{Percentage identity reporting threshold.}

\item{scov}{Query coverage reporting threshold.}

\item{write.new}{\code{logical}. If new alleles found should be written in
\code{dir}.}

\item{dir}{The directory where to put the fasta file written in case a new
allele were found. Not used if \code{write.new = FALSE}.}
}
\value{
The allele number id if an exact match with a reported mlst gene is
found, a name arbitrarily given if a new allele is found, or \code{NA} if no
alleles are found.
}
\description{
Process blastn output. Reports whether the genome
contains a reported allele in pubmlst, or it has a not reported (new)
allele, or if no allele where found. In case a new allele were found, a
fasta file is written.
}
\author{
Ignacio Ferres
}
